open detail screen from recent
